% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calc_ranef}
\alias{calc_ranef}
\title{Calculate random effects based on a subset of variance terms}
\usage{
calc_ranef(ran, model)
}
\arguments{
\item{model}{a model object.}

\item{ran}{a named list of integer vectors representing
  the random effects that should be calculated.}
}
\value{
a numeric vector.
}
\description{
This function calculates random effects of a mixed model based
on a subset of the variance terms.
}
\details{
The names of the list \code{ran} correspond to the names
  of the list returned by \code{getME(model, "cnms")}.

  The integer values within each vector correpsond to the position
  of the name of the random effect.

  If \code{ran} is \code{NULL}, the function will return a vector
  of zeroes.
}
\seealso{
\code{\link{ranef}} for extracting the modes of the
  random effects.
}
\keyword{internal}

