% Generated by roxygen2 (4.0.1): do not edit by hand
\name{effect_labels}
\alias{effect_labels}
\title{Get the labels of the fixed effects of the model}
\usage{
effect_labels(model)
}
\arguments{
\item{model}{a model object.}
}
\value{
a character vector.
}
\description{
The function returns the labels of coefficients of the model.
Note: These labels are not identical to the term labels.
}
\keyword{internal}

