% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_intercept.R
\name{has_intercept}
\alias{has_intercept}
\title{Find out whether a model includes an intercept}
\usage{
has_intercept(model)
}
\arguments{
\item{model}{a model object.}
}
\value{
logical. \code{TRUE} if the intercept is present,
        \code{FALSE} otherwise.
}
\description{
The function tests whether an intercept is present in a model.
}
\keyword{internal}
