% Generated by roxygen2 (4.0.1): do not edit by hand
\name{intercept}
\alias{intercept}
\title{Get the intercept of a model}
\usage{
intercept(model)
}
\arguments{
\item{model}{a model object.}
}
\value{
a length-one numeric vector.
}
\description{
The function returns the value of the beta coeffcient of the model's
intercept.
}
\details{
\code{model} is an object returned by a model fitting
  function (e.g., \code{\link{lm}}, \code{\link{aov}}).

  If the model was fitted without an intercept, the function
  returns \code{0} and generates a warning message.
}
\examples{
require(utils)

data(iris)
fit <- lm(Sepal.Length ~ Sepal.Width * Petal.Length * Petal.Width, data = iris)
summary(fit)
remef:::intercept(fit)
}
\keyword{internal}

