% Generated by roxygen2 (4.0.1): do not edit by hand
\name{keepef}
\alias{keepef}
\title{Keep effects}
\usage{
keepef(model, fix = NULL, ran = NULL, grouping = FALSE,
  keep.intercept = TRUE, inverse = FALSE)
}
\arguments{
\item{model}{a model object.}

\item{fix}{a character of coefficient labels.}

\item{ran}{a named list of integer vectors representing
random effects.}

\item{grouping}{logical. Should lower-order terms associated with the
ones in \code{fix} also be kept? The details are given under
'Details'.}

\item{keep.intercept}{logical. Should the intercept be kept
(default) or be removed?}

\item{inverse}{logical. If \code{TRUE}, the inverse of the link
function is applied (after the effects are removed).}
}
\value{
\code{keepef} returns a numeric vector. The length of the
 vector is equal to the number of observations of the model.
}
\description{
\code{keepef} is used to remove fixed and random effects from repsonse
 variable based on a (linear) mixed-effects model object generated by the
 \pkg{lme4} package.
}
\details{
\code{model} is an object returned by a model fitting
 function of the \pkg{lme4} package, i.e., \code{lmer} for
 linear mixed-effects models or \code{glmer} for generalised
 linear mixed-effects models.
 For more on details on mixed-effects models and partial effects, see
 the help page of \code{\link{partial}}.

 TODO: More details on the funcion and its arguments
}
\examples{
# TODO: some examples
}
\seealso{
\code{\link{remef}} for a wrapper of \code{partial} with
 \code{keep = FALSE}.
}

