% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial.R
\name{ran_as_integer}
\alias{ran_as_integer}
\title{Random-effect names to indices}
\usage{
ran_as_integer(ran, model)
}
\arguments{
\item{ran}{a named list of numeric index vectors or character vectors
(or a mixture of both).}

\item{model}{a model object.}
}
\value{
a named list of numeric index vectors.
}
\description{
Transforms the list of character vectors denoting random effects into a
list of corresponding numeric index vectors.
}
\details{
The function checks the validity of the parameter `ran` for the
  use in the \code{\link{partial}} function. Furthermore, character
  vectors are transformed to index vectors. These operations are based on
  the model fit (parameter \code{model}).
}
\keyword{internal}
