% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ranef_labels}
\alias{ranef_labels}
\title{Get the labels of the random effects of the model.}
\usage{
ranef_labels(model)
}
\arguments{
\item{model}{a model object.}
}
\value{
a named list of character vectors.
}
\description{
The function returns the labels of coefficients of the model.
Note. These labels are not identical to the term labels.
}
\keyword{internal}

