% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/remef-package.R
\docType{package}
\name{remef-package}
\alias{remef-package}
\title{\pkg{remef} removes effects}
\description{
\code{remef} can be used to remove partial effects from variables used in
statistical models.
}
\details{
Mixed-effetcs models include both fixed and random effects. The dependent
variable is expressed as a combination of fixed and random effects and an
error term. \code{remef} is a practical tool for removing any combiantion
of fixed and random effects from the data. The removal of effects is very
helpful for understanding and visualizing certain effects in highly
complex statistical models. The package works with \code{merMod} objects
generated by the package \code{lme4}. For more details on mixed-effects
models and partial effects, have a look at the documentation of the
function \code{\link{partial}}.
}

