% Generated by roxygen2 (4.0.1): do not edit by hand
\name{term_labels}
\alias{term_labels}
\title{Get the labels of the terms of a model.}
\usage{
term_labels(model)
}
\arguments{
\item{model}{a model object.}
}
\value{
a character vector.
}
\description{
The function returns the labels of the terms of the model
Note. The terms are not identical top the coefficients.
}
\keyword{internal}

