% Generated by roxygen2 (4.0.1): do not edit by hand
\name{coef2term}
\alias{coef2term}
\title{Extract the term names of coefficients of a model.}
\usage{
coef2term(ef, model)
}
\arguments{
\item{ef}{a character vector of coefficient labels.}

\item{model}{a model object.}
}
\value{
a character vector of term labels (one element for each
   element in \code{ef}).
}
\description{
The function returns the labels of the terms of the model
Note. The terms are not identical to the coefficients.
}
\details{
\code{model} is an object returned by a model fitting
  function (e.g., \code{\link{lm}}, \code{\link{aov}}).

  The character string passed to \code{ef} must not include the
  intercept (\code{"(Intercept)"}).
  All non-intercept labels of any order are allowed.
}
\examples{
require(utils)

data(iris)
fit <- lm(Sepal.Length ~ poly(Sepal.Width, 3), data = iris)
summary(fit)
coef2term("poly(Sepal.Width, 3)2", fit)

data(warpbreaks)
fit2 <- lm(breaks ~ wool * tension, data = warpbreaks)
summary(fit2)
coef2term(c("woolB", "tensionH"), fit2)
coef2term(c("woolB:tensionM"), fit2)
}
\seealso{
\code{\link{term2coef}} for the inverse function.
}

